requirements.txt
websockets==14.2
google-auth==2.38.0
python-dotenv==1.0.1
requests
google-genai==1.16.1
aiohttp==3.9.1
aiohttp-cors==0.7.0
uvicorn
fastapi
google-cloud-texttospeech
google-auth[pyopenssl]
numpy
scipy